move mount/umount/borgfs command to archiver.mount
